+++ instark/__main__.py
@@ -18,6 +18,7 @@ def main():  # pragma: no cover
     strategy = config['strategy']
 
     resolver = Injectark(strategy=strategy, factory=factory)
+
     Cli(config, resolver).run(sys.argv[1:])
