+++ instark/infrastructure/web/__init__.py
@@ -2,3 +2,4 @@ from gevent import monkey
 monkey.patch_all()  # noqa
 from .base import create_app
 from .server import ServerApplication
+from .hooks import *
