+++ instark/infrastructure/web/base.py
@@ -5,12 +5,12 @@ from .api import create_api
 from .errors import register_error_handler
 
 
-def create_app(config, resolver: Injectark):
+def create_app(config, resolver: Injectark) -> Flask:
     app = Flask(__name__)
     CORS(app)
     app.config.update(config['flask'])
     register_error_handler(app)
-    
+
     create_api(app, resolver)
 
     return app
