+++ instark/infrastructure/web/hooks/__init__.py
@@ -1,0 +1,1 @@
+from .errors import register_error_handler
