+++ instark/infrastructure/web/middleware/authenticate.py
@@ -22,21 +22,22 @@ class Authenticate:
             # self.session_coordinator.set_tenant(tenant_dict)
             authorization = request.headers.get('Authorization', "")
             token = authorization.replace('Bearer ', '')
-            
+
             if not token:
                 token = request.args.get('access_token')
 
             try:
                 token_payload = self.jwt_supplier.decode(
                     token, verify=False)
-                tenant_dict = self.tenant_supplier.get_tenant(token_payload['tid'])
+                tenant_dict = self.tenant_supplier.get_tenant(
+                    token_payload['tid'])
 
                 token_payload = self.jwt_supplier.decode(token, secret=None)
 
                 self.session_coordinator.set_tenant(tenant_dict)
-                
+
                 user_dict = UserSchema().load(token_payload)
-               
+
                 self.session_coordinator.set_user(user_dict)
 
             except Exception as e:
