+++ tests/infrastructure/web/conftest.py
@@ -6,22 +6,21 @@ from datetime import datetime
 from flask.testing import FlaskClient
 from injectark import Injectark
 from instark.infrastructure.core import (
-    build_factory, build_config, Config, JwtSupplier)
+    build_factory, DevelopmentConfig, build_config, Config, JwtSupplier)
 from instark.infrastructure.cli import Cli
-from instark.infrastructure.web import create_app, ServerApplication
+from instark.infrastructure.web import (
+    create_app, ServerApplication, register_error_handler)
 
 
 @fixture
 def app() -> Flask:
-    config = build_config("", os.environ.get('INSTARK_MODE', 'DEV'))
-
+    config = DevelopmentConfig()
     strategy = config['strategy']
     factory = build_factory(config)
 
     resolver = Injectark(strategy=strategy, factory=factory)
 
     app = create_app(config, resolver)
-    app.testing = True
     app = cast(Flask, app.test_client())
 
     return app
@@ -31,8 +30,8 @@ def app() -> Flask:
 def headers() -> dict:
 
     payload_dict = {
-        "tid": "c5934df0-cab9-4660-af14-c95272a92ab7",
-        "uid": "c4e47c69-b7ee-4a06-83bb-b59859478bec",
+        "tid": "1",
+        "uid": "1",
         "name": "John Doe",
         "email": "johndoe@nubark.com",
         "attributes": {},
@@ -44,8 +43,3 @@ def headers() -> dict:
     token = jwt_supplier.encode(payload_dict)
 
     return {"Authorization": (token)}
-
-
-@fixture
-def retrieve_production_conf() -> Config:
-    return build_config("", 'PROD')
